pyproject.toml
[project]
name = "Talk2PowerSystemLLM"
version = "0.0.0a0"
description = "Talk to Power System LLM"
authors = []
readme = "README.adoc"
license = "Apache-2.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "ttyg==1.3.0",
    "ttyg-evaluation==2.2.0",
    "jupyter==1.1.1",
    "langchain-openai==0.3.27",
    "jsonlines==4.0.0",
    "cognite-sdk==7.76.0",
    "pydantic-settings==2.10.1",
    "PyYAML==6.0.2",
    "uvicorn[standard] (==0.35.0)",
    "fastapi==0.116.0",
    "toml==0.10.2",
    "markdown==3.8.2",
]

[project.urls]
repository = "https://github.com/statnett/Talk2PowerSystem_LLM"

[tool.poetry.group.test.dependencies]
pytest = "<9,>=8"
pytest-cov = "<7,>=6"

[tool.poetry.group.test]
optional = true

[[tool.poetry.source]]
name = "statnett"
url = "https://artifactory.statnett.no/artifactory/api/pypi/pypi_statnett_virtual/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
evaluation = 'talk2powersystemllm.scripts.run_evaluation:main'
qa_dataset2rdf = 'talk2powersystemllm.scripts.qa_dataset2rdf:main'
find_min_retrieval_limit_on_dev = 'talk2powersystemllm.scripts.find_min_retrieval_limit_on_dev:main'
